After implementing Detail page for each product with picture
dated 12-05-2016

+++ Flower/src/main/java/service/ProductService.java
@@ -41,7 +41,7 @@ public class ProductService {
     {
     	dao.removeProduct(id);
     }
-    public Long countProduct()
+    public Integer countProduct()
 	{
 		return dao.countProduct();
 	}
